First version of mongodb_shell unit tests
