+++ tests/unit/files/1_mongosh_rs_status.txt
@@ -1,0 +1,115 @@
+{
+  set: 'atlas-j5xqo0-shard-0',
+  date: ISODate("2022-10-01T12:46:58.405Z"),
+  myState: 1,
+  term: Long("187"),
+  syncSourceHost: '',
+  syncSourceId: -1,
+  heartbeatIntervalMillis: Long("2000"),
+  majorityVoteCount: 2,
+  writeMajorityCount: 2,
+  votingMembersCount: 3,
+  writableVotingMembersCount: 3,
+  optimes: {
+    lastCommittedOpTime: { ts: Timestamp({ t: 1664628418, i: 37 }), t: Long("187") },
+    lastCommittedWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+    readConcernMajorityOpTime: { ts: Timestamp({ t: 1664628418, i: 37 }), t: Long("187") },
+    appliedOpTime: { ts: Timestamp({ t: 1664628418, i: 37 }), t: Long("187") },
+    durableOpTime: { ts: Timestamp({ t: 1664628418, i: 37 }), t: Long("187") },
+    lastAppliedWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+    lastDurableWallTime: ISODate("2022-10-01T12:46:58.340Z")
+  },
+  lastStableRecoveryTimestamp: Timestamp({ t: 1664628366, i: 79 }),
+  electionCandidateMetrics: {
+    lastElectionReason: 'stepUpRequestSkipDryRun',
+    lastElectionDate: ISODate("2022-09-30T08:50:21.440Z"),
+    electionTerm: Long("187"),
+    lastCommittedOpTimeAtElection: { ts: Timestamp({ t: 1664527821, i: 36 }), t: Long("186") },
+    lastSeenOpTimeAtElection: { ts: Timestamp({ t: 1664527821, i: 36 }), t: Long("186") },
+    numVotesNeeded: 2,
+    priorityAtElection: 7,
+    electionTimeoutMillis: Long("5000"),
+    priorPrimaryMemberId: 2,
+    numCatchUpOps: Long("0"),
+    newTermStartDate: ISODate("2022-09-30T08:50:21.516Z"),
+    wMajorityWriteAvailabilityDate: ISODate("2022-09-30T08:50:22.805Z")
+  },
+  members: [
+    {
+      _id: 0,
+      name: 'ac-4rp4yqx-shard-00-00.lyyoypw.mongodb.net:27017',
+      health: 1,
+      state: 2,
+      stateStr: 'SECONDARY',
+      uptime: 100724,
+      optime: { ts: Timestamp({ t: 1664628416, i: 44 }), t: Long("187") },
+      optimeDurable: { ts: Timestamp({ t: 1664628416, i: 44 }), t: Long("187") },
+      optimeDate: ISODate("2022-10-01T12:46:56.000Z"),
+      optimeDurableDate: ISODate("2022-10-01T12:46:56.000Z"),
+      lastAppliedWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      lastDurableWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      lastHeartbeat: ISODate("2022-10-01T12:46:56.445Z"),
+      lastHeartbeatRecv: ISODate("2022-10-01T12:46:56.554Z"),
+      pingMs: Long("0"),
+      lastHeartbeatMessage: '',
+      syncSourceHost: 'ac-4rp4yqx-shard-00-01.lyyoypw.mongodb.net:27017',
+      syncSourceId: 1,
+      infoMessage: '',
+      configVersion: 9,
+      configTerm: 187
+    },
+    {
+      _id: 1,
+      name: 'ac-4rp4yqx-shard-00-01.lyyoypw.mongodb.net:27017',
+      health: 1,
+      state: 1,
+      stateStr: 'PRIMARY',
+      uptime: 100758,
+      optime: { ts: Timestamp({ t: 1664628418, i: 37 }), t: Long("187") },
+      optimeDate: ISODate("2022-10-01T12:46:58.000Z"),
+      lastAppliedWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      lastDurableWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      syncSourceHost: '',
+      syncSourceId: -1,
+      infoMessage: '',
+      electionTime: Timestamp({ t: 1664527821, i: 37 }),
+      electionDate: ISODate("2022-09-30T08:50:21.000Z"),
+      configVersion: 9,
+      configTerm: 187,
+      self: true,
+      lastHeartbeatMessage: ''
+    },
+    {
+      _id: 2,
+      name: 'ac-4rp4yqx-shard-00-02.lyyoypw.mongodb.net:27017',
+      health: 1,
+      state: 2,
+      stateStr: 'SECONDARY',
+      uptime: 100454,
+      optime: { ts: Timestamp({ t: 1664628416, i: 55 }), t: Long("187") },
+      optimeDurable: { ts: Timestamp({ t: 1664628416, i: 55 }), t: Long("187") },
+      optimeDate: ISODate("2022-10-01T12:46:56.000Z"),
+      optimeDurableDate: ISODate("2022-10-01T12:46:56.000Z"),
+      lastAppliedWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      lastDurableWallTime: ISODate("2022-10-01T12:46:58.340Z"),
+      lastHeartbeat: ISODate("2022-10-01T12:46:56.663Z"),
+      lastHeartbeatRecv: ISODate("2022-10-01T12:46:56.554Z"),
+      pingMs: Long("1"),
+      lastHeartbeatMessage: '',
+      syncSourceHost: 'ac-4rp4yqx-shard-00-01.lyyoypw.mongodb.net:27017',
+      syncSourceId: 1,
+      infoMessage: '',
+      configVersion: 9,
+      configTerm: 187
+    }
+  ],
+  ok: 1,
+  '$clusterTime': {
+    clusterTime: Timestamp({ t: 1664628418, i: 37 }),
+    signature: {
+      hash: Binary(Buffer.from("e29605ddff2b38c5388a7c27d63be6c98cc5cc35", "hex"), 0),
+      keyId: Long("7097268525600866365")
+    }
+  },
+  operationTime: Timestamp({ t: 1664628418, i: 37 })
+}
