+++ tests/unit/files/2_mongosh_insert.txt
@@ -1,0 +1,5 @@
+DeprecationWarning: Collection.insert() is deprecated. Use insertOne, insertMany, or bulkWrite.
+{
+  acknowledged: true,
+  insertedIds: { '0': ObjectId("63383958c213b3865ee85563") }
+}
