+++ tests/unit/files/3_mongosh_insertOne.txt
@@ -1,0 +1,4 @@
+{
+  acknowledged: true,
+  insertedId: ObjectId("633839a5c213b3865ee85564")
+}
